requirements.txt
# Autonomous Macro Replay System - Dependencies
# Python 3.8+ required

# Core ML & Vision
pillow>=10.0.0

# Embedding API
voyageai>=0.2.0

# Vector Search & Similarity
faiss-cpu>=1.7.4
numpy>=1.24.0
scipy>=1.10.0

# Data Processing
pandas>=2.0.0
tqdm>=4.65.0

# UI & Visualization
streamlit>=1.28.0
plotly>=5.17.0
networkx>=3.1
matplotlib>=3.7.0

# Utilities
pyyaml>=6.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0

# Development & Testing
jupyter>=1.0.0
ipykernel>=6.25.0
pytest>=7.4.0


